+++ index.d.ts
@@ -35,7 +35,7 @@ declare module '@lokalise/node-api' {
     original_filenames?: boolean;
     bundle_structure?: string;
     directory_prefix?: string;
-    all_platforms?: boolean
+    all_platforms?: boolean;
     filter_langs?: any[];
     filter_data?: any[];
     filter_filenames?: any[];
@@ -124,6 +124,18 @@ declare module '@lokalise/node-api' {
     translation_tier_name: string;
     briefing: string;
     total: number;
+  }
+
+  export interface QueuedProcess {
+    process_id: string;
+    type: string;
+    status: string;
+    message: string;
+    created_by: number;
+    created_by_email: string;
+    created_at: string;
+    created_at_timestamp: number;
+    details?: object[];
   }
 
   export interface PaymentCard {
@@ -262,6 +274,7 @@ declare module '@lokalise/node-api' {
   export interface UploadFileParams {
     data: string;
     filename: string;
+    queue?: boolean;
     lang_iso: string;
     convert_placeholder?: boolean;
     detect_icu_plurals?: boolean;
@@ -316,13 +329,16 @@ declare module '@lokalise/node-api' {
     protected static endpoint: string | null;
     protected static prefixURI: string | null;
     protected static elementClass: any | null;
+    protected static secondaryElementNameSingular: string | null;
+    protected static secondaryElementClass: any;
     get(id: any, params?: StandartParams, body?: any): Promise<any>;
     list(params?: StandartParams): Promise<any[]>;
     create(body: any, params?: StandartParams): Promise<any>;
     update(id: any, body: any, params?: StandartParams): Promise<any>;
     delete(id: any, params?: StandartParams): Promise<any>;
     protected populateObjectFromJsonRoot(json: Object): this;
-    protected populateObjectFromJson(json: Object): this;
+    protected populateSecondaryObjectFromJsonRoot(json: any): this;
+    protected populateObjectFromJson(json: Object, secondary: boolean): this;
     protected populateArrayFromJson(json: Array<any>): this[];
     protected returnBareJSON(json: any): any;
     protected handleReject(data: any): void;
@@ -342,8 +358,8 @@ declare module '@lokalise/node-api' {
     protected static prefixURI: string;
     protected static elementClass: Object;
     create(body: any, params: StandartParams): Promise<any>;
-    update(id: any, body: any, params : StandartParams) : Promise<any>;
-    merge(id: any, body: any, params: StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
+    merge(id: any, body: any, params: StandartParams): Promise<any>;
   }
 
   export class Comments extends BaseCollection {
@@ -367,6 +383,8 @@ declare module '@lokalise/node-api' {
     protected static rootElementName: string;
     protected static prefixURI: string;
     protected static elementClass: Object;
+    protected static secondaryElementNameSingular: string;
+    protected static secondaryElementClass: Object;
     list(params?: FileParams): Promise<this[]>;
     upload(project_id: string, upload: UploadFileParams): Promise<any>;
     download(project_id: string, download: DownloadFileParams): Promise<any>;
@@ -416,19 +434,26 @@ declare module '@lokalise/node-api' {
     empty(project_id: any): void;
   }
 
+  export class QueuedProcesses extends BaseCollection {
+    protected static rootElementName: string;
+    protected static rootElementNameSingular: string;
+    protected static prefixURI: string;
+    protected static elementClass: Object;
+  }
+
   export class Screenshots extends BaseCollection {
     protected static rootElementName: string;
     protected static rootElementNameSingular: string;
     protected static prefixURI: string;
     protected static elementClass: Object;
     create(raw_body: any, params: StandartParams): Promise<any>;
-    update(id: any, body: any, params : StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
   }
 
   export class Snapshots extends BaseCollection {
-    protected static rootElementName:string;
+    protected static rootElementName: string;
     protected static rootElementNameSingular: string;
-    protected static prefixURI:string;
+    protected static prefixURI: string;
     protected static elementClass: Object;
     restore(id: any, params: StandartParams): Promise<any>;
     create(body: any, params: StandartParams): Promise<any>;
@@ -441,7 +466,7 @@ declare module '@lokalise/node-api' {
     protected static elementClass: Object;
 
     create(body: any, params: StandartParams): Promise<any>;
-    update(id: any, body: any, params : StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
   }
 
   export class TeamUsers extends BaseCollection {
@@ -449,7 +474,7 @@ declare module '@lokalise/node-api' {
     protected static rootElementNameSingular: string;
     protected static prefixURI: string;
     protected static elementClass: Object;
-    update(id: any, body: any, params : StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
   }
 
   export class Teams extends BaseCollection {
@@ -469,7 +494,7 @@ declare module '@lokalise/node-api' {
     protected static rootElementNameSingular: string;
     protected static prefixURI: string;
     protected static elementClass: Object;
-    update(id: any, body: any, params : StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
   }
 
   export class TranslationStatuses extends BaseCollection {
@@ -488,7 +513,7 @@ declare module '@lokalise/node-api' {
     protected static prefixURI: string;
     protected static elementClass: Object;
     create(body: any, params?: StandartParams): Promise<any>;
-    update(id: any, body: any, params : StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
     add_projects_to_group(team_id: any, group_id: any, raw_body: any[], params: any): Promise<any>;
     remove_projects_from_group(team_id: any, group_id: any, raw_body: any[], params: any): Promise<any>;
     add_members_to_group(team_id: any, group_id: any, raw_body: any[], params: any): Promise<any>;
@@ -502,8 +527,8 @@ declare module '@lokalise/node-api' {
     protected static elementClass: Object;
 
     create(body: any, params: StandartParams): Promise<any>;
-    update(id: any, body: any, params: StandartParams) : Promise<any>;
-    regenerate_secret(id: any, params : StandartParams) : Promise<any>;
+    update(id: any, body: any, params: StandartParams): Promise<any>;
+    regenerate_secret(id: any, params: StandartParams): Promise<any>;
   }
 
   export class LocaliseApiMethods {
@@ -530,7 +555,6 @@ declare module '@lokalise/node-api' {
 
   export class LokaliseApi extends LocaliseApiMethods {
     static apiKey: string | null;
-    private static _instance: LokaliseApi;
     apiKey: string;
     /**
      * Instantiate LokaliseApi to have access to methods