+++ src/collections/base_collection.ts
@@ -8,15 +8,20 @@ export class BaseCollection {
   protected static endpoint: string | null = null;
   protected static prefixURI: string | null = null;
   protected static elementClass: any = null;
+  
+  // Secondaries are used when an instance of a different class has to be created
+  // For example, uploading a File may return a QueuedProcess
+  protected static secondaryElementNameSingular: string | null = null;
+  protected static secondaryElementClass: any = null;
 
   // Workaround for handling HTTP header pagination params
 
   totalResults: number | null  = null;
   totalPages: number | null  = null;
   resultsPerPage: number | null = null;
-  currentPage:number | null = null;
+  currentPage: number | null = null;
 
-  get(id: any, params: StandartParams = {}, body: any = null) : Promise<any> {
+  get(id: any, params: StandartParams = {}, body: any = null): Promise<any> {
     params['id'] = id;
     return this.createPromise('GET', params, this.populateObjectFromJsonRoot, this.handleReject, body)
   }
@@ -29,41 +34,54 @@ export class BaseCollection {
     return this.createPromise('POST', params, this.populateObjectFromJson, this.handleReject, body);
   }
 
-  update(id: any, body: any, params: StandartParams = {}) : Promise<any> {
+  update(id: any, body: any, params: StandartParams = {}): Promise<any> {
     params['id'] = id;
     return this.createPromise('PUT', params, this.populateObjectFromJson, this.handleReject, body);
   }
 
-  delete(id: any, params: StandartParams = {}) {
+  delete(id: any, params: StandartParams = {}): Promise<any> {
     params['id'] = id;
     return this.createPromise('DELETE', params, this.returnBareJSON, this.handleReject, null);
   }
 
-  populatePaginationDataFor(headers:any) {
+  populatePaginationDataFor(headers: any): void {
     this.totalResults = parseInt(headers['x-pagination-total-count']);
     this.totalPages = parseInt(headers['x-pagination-page-count']);
     this.resultsPerPage = parseInt(headers['x-pagination-limit']);
     this.currentPage = parseInt(headers['x-pagination-page']);
+    return;
   }
 
   protected populateObjectFromJsonRoot(json: any): this {
-    let childClass = <typeof BaseCollection>this.constructor;
+    const childClass = <typeof BaseCollection>this.constructor;
     if (childClass.rootElementNameSingular != null) {
       json = json[childClass.rootElementNameSingular];
     }
     return this.populateObjectFromJson(json);
   }
 
-  protected populateObjectFromJson(json: Object): this {
-    let childClass = <typeof BaseCollection>this.constructor;
-    return new childClass.elementClass(json);
+  protected populateSecondaryObjectFromJsonRoot(json: any): this {
+    const childClass = <typeof BaseCollection>this.constructor;
+    if (childClass.secondaryElementNameSingular != null) {
+      json = json[childClass.secondaryElementNameSingular];
+    }
+    return this.populateObjectFromJson(json, true);
+  }
+
+  protected populateObjectFromJson(json: Object, secondary: boolean = false): this {
+    const childClass = <typeof BaseCollection>this.constructor;
+    if (secondary) {
+      return new childClass.secondaryElementClass(json);
+    } else {
+      return new childClass.elementClass(json);
+    }
   }
 
   protected populateArrayFromJson(json: Array<any>): this[] {
-    let childClass = <typeof BaseCollection>this.constructor;
-    let arr: this[] = [];
-    let jsonArray = json[(childClass as any).rootElementName];
-    for (let obj of jsonArray) {
+    const childClass = <typeof BaseCollection>this.constructor;
+    const arr: this[] = [];
+    const jsonArray = json[(<any>childClass).rootElementName];
+    for (const obj of jsonArray) {
       arr.push(this.populateObjectFromJson(obj));
     }
     return arr;
@@ -81,17 +99,17 @@ export class BaseCollection {
     return this.populateApiErrorFromJson(data);
   }
 
-  protected createPromise(method: any, params: any, resolveFn: any, rejectFn = this.handleReject, body: any = null, uri: any = null) : Promise<any> {
-    let childClass = <typeof BaseCollection>this.constructor;
+  protected createPromise(method: any, params: any, resolveFn: any, rejectFn = this.handleReject, body: any = null, uri: any = null): Promise<any> {
+    const childClass = <typeof BaseCollection>this.constructor;
     if (uri == null) {
       uri = childClass.prefixURI;
     }
     return new Promise((resolve, reject) => {
-      let response: ApiRequest = new ApiRequest(uri, method, body, params);
+      const response: ApiRequest = new ApiRequest(uri, method, body, params);
       response.promise.then((result) => {
-        let headers = result['headers'];
+        const headers = result['headers'];
         this.populatePaginationDataFor(headers);
-        let json = result['body'];
+        const json = result['body'];
         resolve(resolveFn.call(this, json));
       }).catch((data) => {
         reject(rejectFn.call(this, data));