+++ src/interfaces/standart_params.ts
@@ -1,5 +1,5 @@
 export interface StandartParams {
   page?: number;
   limit?: number;
-  [paramName: string] : any;
+  [paramName: string]: any;
 }