+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Lokalise API Node Client Changelog
 
-## Unreleased
+## 4.0.0 (14-May-20)
 
+* **Breaking change** [All translation files are now uploaded in the background](https://github.com/lokalise/node-lokalise-api#upload-a-file). A `QueuedProcess` with the job status will be returned as a result. Synchronous uploading is still supported by the API but will be removed in the near future (use version 3 to upload synchronously).
+* Added support for [`QueuedProcess` endpoint](https://app.lokalise.com/api2docs/curl/#resource-queued-processes)
 * Test against Node 14
 * Fix `webhook_id` type for `Webhooks` endpoint
+* Added linter
 
 ## 3.0.3 (22-Apr-20)